WinDV.rc
Copying this file over, as this is what I hope to convert to JSX. I am more interested in learning about it as a format, rather than solely only converting it by hand. Gotta get the full depth of knowledge here to figure out how it works, and why everyone chooses this on Windows. Why does it have to be Windows-only? I want to know why!

https://github.com/hfiggs/WinDV
https://github.com/Offroaders123/WinDV