Initial Porting
Just doing this by hand. Might ask for GPT to help out. I like learning about Python too though, so I don't mind learning more about the language and how you structure things.

Getting sleepy. Would keep working on it, but I have to work tomorrow and it's already late, and I'm sleepy now.

I want to make a programming album, thought of this concept a few years ago. Each song is the method of a class. One might be Constructor. It is kind of like a reference to Distruptr, and Deconstruction. It sounds cool too.